django/requirements.txt
Django==1.11.7
djangorestframework==3.7.3
django-filter==1.1.0
Markdown==2.6.9
